Completed Module JobApplied,Events,EventJoined,Friends DTO,DAO,DAOImpl,TestCase all completed and working fine

+++ collaboration/src/main/java/com/sbkchat/collaboration/dao/FriendsDAO.java
@@ -17,21 +17,21 @@ public interface FriendsDAO {
 	boolean deleteFriend(Friends friends);
 	
 	// get friend
-	Friends getFriend(Long id);
+	Friends getFriend(int id);
 	
 	// list of friends
 	List<Friends> list();
 	
 	// list of friends by friend id
-	List<Friends> list(Long id);
+	List<Friends> list(int id);
 	
 	// list of friends by status
 	List<Friends> list(String status);
 	
 	// list of user  for getting list of my friends 
-	List<User> myFriends(Long id);
+	List<User> myFriends(int id);
 	
 	// list of user for getting list that has no  friends for me
-	List<User> noFriends(Long id);
+	List<User> noFriends(int id);
 	
 }
